+++ Rebus.Tests/Transport/Msmq/MsmqTransportFactory.cs
@@ -1,6 +1,5 @@
-﻿using System;
+using System;
 using System.Collections.Generic;
-using NUnit.Framework;
 using Rebus.Extensions;
 using Rebus.Tests.Contracts.Transports;
 using Rebus.Transport;
@@ -8,12 +7,6 @@ using Rebus.Transport.Msmq;
 
 namespace Rebus.Tests.Transport.Msmq
 {
-    [TestFixture, Category(Categories.Msmq)]
-    public class MsmqBasicSendReceive : BasicSendReceive<MsmqTransportFactory> { }
-
-    [TestFixture, Category(Categories.Msmq)]
-    public class MsmqMessageExpiration : MessageExpiration<MsmqTransportFactory> { }
-
     public class MsmqTransportFactory : ITransportFactory
     {
         readonly List<IDisposable> _disposables = new List<IDisposable>();